+++ src/components/banner/banner.js
@@ -1,7 +1,7 @@
-import React, { useEffect, useState , useCallback} from "react"
+import React, { useEffect, useState, useCallback } from "react"
 import { Link } from "gatsby"
 
-import { ContentBanner, BoxCenter, WrapperBoxCenter, BoxBottom , Dots , BannerImage} from "./styled"
+import { ContentBanner, BoxCenter, WrapperBoxCenter, BoxBottom, Dots, BannerImage } from "./styled"
 import { Container, FlexList } from "../../shared/styled-shared"
 
 import icon1 from "../../assets/imgs/icon1.png"
@@ -13,12 +13,12 @@ import { graphql, useStaticQuery } from "gatsby"
 
 
 export default function Banner() {
-      
-    const [ image, setImage ] = useState('');
-    const [ listImages, setListImages ] = useState([]);
+
+  const [image, setImage] = useState('');
+  const [listImages, setListImages] = useState([]);
 
 
-      const data = useStaticQuery(graphql`
+  const data = useStaticQuery(graphql`
       query Banners {
         images:allFile(filter:{
           relativeDirectory:{
@@ -40,55 +40,55 @@ export default function Banner() {
     `)
 
   const dataImages = () => {
-    const images = data.images.nodes.reduce((reduce,item,index) => {
-      if(item.childImageSharp){
+    const images = data.images.nodes.reduce((reduce, item, index) => {
+      if (item.childImageSharp) {
         reduce.push({
-          id:index,
+          id: index,
           src: (window.innerWidth <= 800) ? item.childImageSharp.fixed.src : item.childImageSharp.original.src
         })
       }
       return reduce;
-    },[]);
+    }, []);
     setListImages(images);
     setImage(images[0])
   }
-   
+
   useEffect(() => {
     dataImages();
-    const event = window.addEventListener('resize',dataImages)
-    return () => window.removeEventListener('resize',dataImages);
-  },[]);
-  
+    const event = window.addEventListener('resize', dataImages)
+    return () => window.removeEventListener('resize', dataImages);
+  }, []);
+
   const next = useCallback(() => {
     const numberOfImages = listImages.length;
-    if(!numberOfImages) return;
+    if (!numberOfImages) return;
     const indexOf = listImages.indexOf(image);
     const nextImage = indexOf + 1;
-    if(listImages[nextImage] !== undefined)
+    if (listImages[nextImage] !== undefined)
       return setImage(listImages[nextImage])
 
-      return setImage(listImages[0])
-  },[listImages,image])
+    return setImage(listImages[0])
+  }, [listImages, image])
 
   const prev = useCallback(() => {
     const numberOfImages = listImages.length;
-    if(!numberOfImages) return;
+    if (!numberOfImages) return;
     const indexOf = listImages.indexOf(image);
     const nextImage = indexOf - 1;
-    if(listImages[nextImage] !== undefined){
+    if (listImages[nextImage] !== undefined) {
       return setImage(listImages[nextImage])
-    }else{
+    } else {
       return setImage(listImages[numberOfImages - 1])
     }
-  },[listImages,image])
+  }, [listImages, image])
 
   useEffect(() => {
     const slideOn = setInterval(() => {
       next();
-    },3000);
+    }, 3000);
     return () => clearInterval(slideOn);
-  },[])
-      
+  }, [])
+
   return (
     <ContentBannerImage images={listImages} active={image}>
       <Container>
@@ -153,40 +153,40 @@ export default function Banner() {
           </BoxBottom>
         </WrapperBoxCenter>
         <Dots>
-          <FlexList flow="row" style={{justifyContent:'space-between', alignItems:'center'}}>
-                <li onClick={() => prev()}>
-                  <svg width="8" height="12" viewBox="0 0 8 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-                  <path d="M7.74347 1.41L3.16347 6L7.74347 10.59L6.33347 12L0.333469 6L6.33347 4.4819e-08L7.74347 1.41Z" fill="white"/>
+          <FlexList flow="row" style={{ justifyContent: 'space-between', alignItems: 'center' }}>
+            <li onClick={() => prev()}>
+              <svg width="8" height="12" viewBox="0 0 8 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+                <path d="M7.74347 1.41L3.16347 6L7.74347 10.59L6.33347 12L0.333469 6L6.33347 4.4819e-08L7.74347 1.41Z" fill="white" />
+              </svg>
+            </li>
+            {
+              listImages.map((dot, index) => (
+                <li key={index} className={(image.id === dot.id) ? 'active' : ''} onClick={() => setImage(listImages[index])}>
+                  <svg width="5" height="5" viewBox="0 0 5 5" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <circle cx="2.03845" cy="2.29492" r="2" transform="rotate(-90 2.03845 2.29492)" fill="white" />
                   </svg>
                 </li>
-                {
-                  listImages.map( (dot,index) => (
-                    <li key={index} className={(image.id === dot.id) ? 'active' : ''} onClick={() => setImage(listImages[index])}>
-                      <svg width="5" height="5" viewBox="0 0 5 5" fill="none" xmlns="http://www.w3.org/2000/svg">
-                          <circle cx="2.03845" cy="2.29492" r="2" transform="rotate(-90 2.03845 2.29492)" fill="white"/>
-                      </svg>
-                    </li>
-                  ))
-                }
-                <li onClick={() => next()}>
-                  <svg width="8" height="12" viewBox="0 0 8 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-                    <path d="M0.333496 10.59L4.9135 6L0.333496 1.41L1.7435 0L7.7435 6L1.7435 12L0.333496 10.59Z" fill="white"/>
-                  </svg>
-              </li>
-            </FlexList>
-      </Dots>
+              ))
+            }
+            <li onClick={() => next()}>
+              <svg width="8" height="12" viewBox="0 0 8 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+                <path d="M0.333496 10.59L4.9135 6L0.333496 1.41L1.7435 0L7.7435 6L1.7435 12L0.333496 10.59Z" fill="white" />
+              </svg>
+            </li>
+          </FlexList>
+        </Dots>
       </Container>
     </ContentBannerImage>
   )
 }
 
-function ContentBannerImage({children,images,active}) {
+function ContentBannerImage({ children, images, active }) {
   return (<ContentBanner>
-      {
-        images.map(item => <BannerImage className={(active.id === item.id) ? 'active' : 'inactive'} key={item.id} src={item.src} alt="banner"/>)
-      }
-      <div>
-        {children}
-      </div>
+    {
+      images.map(item => <BannerImage className={(active.id === item.id) ? 'active' : 'inactive'} key={item.id} src={item.src} alt="banner" />)
+    }
+    <div>
+      {children}
+    </div>
   </ContentBanner>)
 }
